fix(signup): move files for page to component
